+++ ExtendedCheats/Overrides.cs
@@ -2,7 +2,6 @@
 using BepInEx.Configuration;
 using SimpleJSON;
 using System;
-using System.Collections.Generic;
 using System.IO;
 using UnityEngine;