+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Convenience
 - 1.0    Initial release
 - 1.0.1  Fix audio turning off permanently in some situations
+- 1.0.2  Fix volume getting maxed out sometimes
 
 # DevKillsList
 - 1.0    Initial release
@@ -12,6 +13,6 @@
          Revamp overrides
          Actually enable multi-knockout corpse throws
 - 1.1.1  Fix startingAmmo and maxHealth reset
-- curr   Add scene statistics reset command
+- 1.2    Add scene statistics reset command
          Add in-game showdown mode pause toggle
          Allow disabling skills and characters